# server/docs/swagger.yaml
basePath: /api
definitions:
  github_com_ai-comic-generator_server_internal_common.BaseResponse:
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        example: 0
        type: integer
      data:
        description: 响应数据，成功时返回；失败时通常为 null
      message:
        description: 响应消息，成功时为 ok，失败时为错误描述
        example: ok
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.CustomComicInfo:
    properties:
      aspectRatio:
        type: string
      createTime:
        type: string
      errorMessage:
        type: string
      id:
        type: integer
      imageBackend:
        type: string
      panelCount:
        type: integer
      panelImages:
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.PanelImageResult'
        type: array
      prompt:
        type: string
      referenceImages:
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.ReferenceImage'
        type: array
      status:
        type: string
      taskId:
        type: string
      updateTime:
        type: string
      userAccount:
        type: string
      userId:
        type: integer
      userName:
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.CustomComicPageResult:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      records:
        items:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.CustomComicInfo'
        type: array
      total:
        type: integer
    type: object
...
  github_com_ai-comic-generator_server_internal_model.PanelImageResult:
    properties:
      imagePrompt:
        description: 实际使用的生图 Prompt（便于追溯）
        type: string
      method:
        description: 图片来源：AI 生成 或 用户上传
        enum:
        - AI_GENERATE
        - UPLOAD
        type: string
      panelNo:
        description: 对应分镜格序号
        type: integer
      url:
        description: 图片访问地址
        type: string
    type: object
...
  github_com_ai-comic-generator_server_internal_model.QueryCustomComicRequest:
    properties:
      pageNum:
        example: 1
        type: integer
      pageSize:
        example: 10
        type: integer
      status:
        enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        type: string
      userId:
        type: integer
    type: object
...
  github_com_ai-comic-generator_server_internal_model.ReferenceImage:
    properties:
      index:
        type: integer
      name:
        type: string
      url:
        type: string
    type: object
  github_com_ai-comic-generator_server_internal_model.RegenerateCustomPanelRequest:
    properties:
      panelNo:
        type: integer
      prompt:
        description: 可选：覆盖该格生图提示；空则沿用已有 imagePrompt 或从大纲重解析
        type: string
      taskId:
        type: string
    required:
    - panelNo
    - taskId
    type: object
...
info:
  contact: {}
  description: REST API。HTTP 多为 200，业务结果看 JSON 的 code（0 成功）；鉴权接口需 Cookie session。
  title: AI Comic Generator API
  version: "1.0"
paths:
...
  /comic/custom/create:
    post:
      consumes:
      - application/json
      - multipart/form-data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 创建自定义漫画
      tags:
      - 自定义漫画
  /comic/custom/download:
    get:
      parameters:
      - description: 任务 ID
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP 文件
          schema:
            type: file
      security:
      - SessionCookie: []
      summary: 下载分镜 ZIP
      tags:
      - 自定义漫画
  /comic/custom/get:
    get:
      parameters:
      - description: 任务 ID
        in: query
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 查询自定义任务
      tags:
      - 自定义漫画
  /comic/custom/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页与筛选
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.QueryCustomComicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.CustomComicPageResult'
              type: object
      security:
      - SessionCookie: []
      summary: 自定义任务分页
      tags:
      - 自定义漫画
  /comic/custom/regenerate-panel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务与分镜序号
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ai-comic-generator_server_internal_model.RegenerateCustomPanelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ai-comic-generator_server_internal_common.BaseResponse'
      security:
      - SessionCookie: []
      summary: 自定义单格重绘
      tags:
      - 自定义漫画
...
schemes:
- http
- https
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"